More refactoring: removed the split between ''render'' and ''display'' phases imposed somewhat artificially by {{{trac.Module}}}. Modules now need to check the ''format'' parameter themselves, and change their rendering accordingly.
Also, the special format ''hdf'' is no longer supported for dumping the HDF dataset: Instead you can now use the query-string parameter ''hdfdump=1'' (the value doesn't matter) to get a HDF dump of the page, independent of the format. So it's now possible to get the HDF dump of the Timeline RSS page, and not only the Timeline HTML representation, for example.


git-svn-id: https://svn.edgewall.org/repos/trac/trunk@1401 af82e41b-90c4-0310-8c96-b1721e28e2e2