''xref branch'': the ticket dependencies feature (from #31) is implemented.
Ticket dependencies are implemented using __depends-on__ relations.
In particular, circular dependencies are prevented and it's not 
possible to close a ticket until all the tickets it depends on 
are closed.

The __depends-on__ relations are created for all references found 
in the wiki markup entered in a __depends-on__ textarea.
This allows to give some context and nuance to the dependency.

e.g.:
{{{
  First thing, do #3.
  Also, doing #4 before would help
}}}

Note that this context is currently correctly recorded in the {{{xref}}} table, 
but the wiki text itself should also be recorded in a specific ticket facet.
This will be done later, as it involves a schema change.

Also in this changeset:
 * new {{{relation}}} cs macro for displaying relations (they are displayed underlined now)
 * new {{{plural}}} cs macro for selecting a text based 
   on whether a quantity is superior to one or not

git-svn-id: https://svn.edgewall.org/repos/trac/branches/cboos-dev@1412 af82e41b-90c4-0310-8c96-b1721e28e2e2