''xref-branch'': basic objectification done.
There's now an {{{object_factory}}} modeled after the {{{module_factory}}}.
The !WikiPage and Ticket classes are now also inheriting !TracObj.

Small simplification of the presentation of cross-references.

From this point, I'll create the ''trac-obj-branch'' to go on
with the more advanced stuff about objectification of Trac,
and the ''xref-branch'' will start its (hopefully brief!) way
towards acceptance in ''trunk''.


git-svn-id: https://svn.edgewall.org/repos/trac/branches/cboos-dev@1486 af82e41b-90c4-0310-8c96-b1721e28e2e2