''category-branch'': implemented ticket categories (see #919 and #1399).
This requires a database upgrade, which creates by default 3 categories:
 * Bug
 * Feature
 * Task

It splits the existing tickets between Bugs and Features based on the ''severity''
information: an ''enhancement'' will be converted to a ''Feature'' having a 
''normal'' severity.

If later the categories are all deleted (with the trac-admin), then Trac will
behave essentially as without this patch.

Feedback welcomed, as usual.


git-svn-id: https://svn.edgewall.org/repos/trac/branches/cboos-dev@1489 af82e41b-90c4-0310-8c96-b1721e28e2e2