InterTrac: intertrac links are implemented.
A little refactoring of the module link regexp was also done along the way:
it is now built using the list of modules.

InterTrac keys of arbitrary length are allowed, in case they're needed, 
but 1-letter keys should be enough 99% of the time.

The short-hand syntax is available for ticket (e.g. #T234) and
changesets (e.g. [T1505]).

The TracIni configuration for intertrac links is done
in a very simple way, here's an example for the key 'T'
(or 't', keys are case insensitive):

{{{
[intertrac]
t.title = Edgewall's Trac for Trac
t.trac = http://projects.edgewall.com/trac
t.svn = http://svn.edgewall.com/repos/trac
}}}

(The .svn property is not required, but might be used later
for {{{svn:external}}} support)


git-svn-id: https://svn.edgewall.org/repos/trac/branches/cboos-dev@1505 af82e41b-90c4-0310-8c96-b1721e28e2e2