Reworked the special handling of newlines in commit messages.
Newlines in commit messages have to be preserved
(see #352 and #124), but the initial implementation 
(in [776]) did it in a way that either was not working 
an all platforms, or introduced side-effects (see #1181 and #1206).

The new implementation does the escaping at the !WikiFormatter
level, only on lines that actually need to be escaped.

Also, the {{{shorten_line}}} utility has been rewritten
in order to "pack in" more information in the given number
of characters.
This is useful with changeset comments with an empty new line
or no useful first line, like
{{{
fixes:
  * the first fix...
  * the second fix...
}}}


git-svn-id: https://svn.edgewall.org/repos/trac/trunk@1561 af82e41b-90c4-0310-8c96-b1721e28e2e2