When Trac uses SQLite3/PySQLite2 for its DatabaseBackend, concurrent writes were not handled well: this could lead to persistent database locks and, occasionally, to crashes (see #2170).
With the present changes, the issue is fixed, ''provided'' the SQLite library
which is used was compiled in thread-safe mode. 
This is always the case on Windows, but Linux users (and probably all Unix users)
will certainly have to recompile their SQLite package (3.2.7 preferred) 
using the `--enable-threadsafe` configuration option.

Note that ''without'' a threadsafe SQLite, Trac ''will'' continue to work,
simply it will still be vulnerable to the abovementioned problem.

Finally, note that ''without'' this patch, Trac simply won't work with
a threadsafe SQLite binary on Linux (see #2196). 


git-svn-id: https://svn.edgewall.org/repos/trac/trunk@2355 af82e41b-90c4-0310-8c96-b1721e28e2e2