DatabaseBackend refactoring: the old db.py file has been split in several files now placed in the trac/db subfolder.
Also, a new `IDatabaseBackend` interface has been defined, 
and this is used for better dispatching from the database connection string
to the appropriate backend.

This greater flexibility would allow to use the pysqlite1 binding
for one environment and the pysqlite2 binding for another, for example.


While the changes were primarily done for cleaning up the current implementation, 
this will also make it easier to add new database backends.



git-svn-id: https://svn.edgewall.org/repos/trac/branches/cboos-dev@2371 af82e41b-90c4-0310-8c96-b1721e28e2e2