vc-refactoring: `IScmBackendInterface` and `ScmBackendManager` introduced.
This is similar to what I did for the source:branches/cboos-dev/db-refactoring
so there will be some opportunity for code sharing.

Now there is support for creating a Trac environment without specifying a repository
(see #156).

Also, by specifying the prefix 'direct-svn-fs', one can use the Subversion backend
without the cache, which can be useful for development or if you're annoyed by the
'resync' command...


git-svn-id: https://svn.edgewall.org/repos/trac/branches/cboos-dev@2464 af82e41b-90c4-0310-8c96-b1721e28e2e2