vc-refactoring: The versioncontrol.web_ui modules are now abstract components.
This makes it possible to slightly change their behavior in derived classes 
for supporting other SCM backends.

'''Note: this is most probably a temporary situation, if the coming up changes
are propagated to the generic classes, there will be no need for that.'''

Also, the ''Next Changeset'' link is sometimes not existing, even if 
the changeset currently shown is not the last one (i.e. end of branch situation).




git-svn-id: https://svn.edgewall.org/repos/trac/branches/cboos-dev@2465 af82e41b-90c4-0310-8c96-b1721e28e2e2