In the changeset information, the EDIT changes now contain the exact previous location and previous revision of the base.
Though this what not a real problem, as Subversion was nevertheless able to find the 
correct old content, this provides more useful information in the changeset view.

A `resync` operation is needed to take benefit from this change for old changesets.

Thanks to Antoine Delvaux for noticing and reporting the issue.



git-svn-id: https://svn.edgewall.org/repos/trac/trunk@2562 af82e41b-90c4-0310-8c96-b1721e28e2e2