Refactoring of the Mimeview HTML Preview.
The actual content to be previewed is only retrieved and eventually
converted to unicode when it's required.
For example, when previewing images, the content of the image will not
be read anymore, only the url will be used.

In the various `render` methods, the `content` data might be a `str` 
object, an `unicode` object, or an object that can be `read` to obtain
one of the former.

I did a lot of documentation updates in order to make this clear.
Under normal circumstances, `IHTMLPreviewRenderer`s will receive
a `str` object when they deal with data coming from the repository
or from attachment files, and they'll receive an `unicode` object
when they receive data from Wiki "code" blocks.

The name of the 4^th^ argument of the `Mimeview.render` method was
changed to be an `url` instead of a `detail` or `rev`.
This decouples the `mimeview` library from the `versioncontrol` library.



git-svn-id: https://svn.edgewall.org/repos/trac/trunk@3124 af82e41b-90c4-0310-8c96-b1721e28e2e2