Mimeview fixes, following-up to [3124].
The Mimeview.render should also handle `content` arguments which are objects with a `read` method. If the `mimetype` argument was not given, auto-detection would have failed.
Here, we take care to read the actual content, if needed, and at most once.

The logic about expanding tabs in the content has also been enhanced a bit, 
so that not only the tab expansion is done at most once, but also the 
expanded content is only given to the `IHTMLPreviewRenderer` who actually
requires an expanded content.

git-svn-id: https://svn.edgewall.org/repos/trac/trunk@3179 af82e41b-90c4-0310-8c96-b1721e28e2e2