Fixed detection of changes in the Session object.
The order of the `items()` of two equivalent dictionaries is not necessarily the same if those dictionaries themselves contain dictionaries.

In our case, this resulted in a write into the db for ''every'' `Session.save()` call, even if there were no actual change, as soon as the user had a custom query saved in his session, as we store a dictionary under the 'query_constraints' key.

Probably one reason explaining the high rate of database locks we're seeing (see #3446).


git-svn-id: https://svn.edgewall.org/repos/trac/trunk@3616 af82e41b-90c4-0310-8c96-b1721e28e2e2