TracWiki: implement simple PageTemplates (#962)
The default content for a new wiki page can be chosen from a list of page templates.

This list is made up from the wiki pages having a name starting with ''PageTemplates/''.
The initial content of a new page will simply be the content of the chosen template page, or a blank page if the special ''(blank page)'' entry is selected. When there's actually no wiki pages matching that prefix, the initial content will always be the blank page and the list selector will not be shown (i.e. this matches the behavior we had up to now).

To create a new template, simply create a new page having a name starting with ''PageTemplates/''.

(Hint: one could even create a ''!PageTemplates/Template'' for facilitating the creation of new templates!)

git-svn-id: https://svn.edgewall.org/repos/trac/trunk@3881 af82e41b-90c4-0310-8c96-b1721e28e2e2