Added a `raw-attachment:<filename>:<type>:<id>` TracLinks for getting access to attached files using the raw format.
This goes together with the `/raw-attachment` URL prefix, which can be substituted to `/attachment` to get direct access to the content of the attachment. This kind of URLs are now used by the ''Original Format'' alternate download links.

This interacts better with IExplorer and Firefox extensions such as ''PDF Download'' (fixes #2974).

git-svn-id: https://svn.edgewall.org/repos/trac/trunk@4246 af82e41b-90c4-0310-8c96-b1721e28e2e2