Add a `timeline:` TracLinks, for creating links to the Timeline from within wiki text.
The format to use for the link is ISO:8601. If the timezone is not specified, the timezone is assumed to be UTC. At least the year and the month has to be specified. If the hour is specified, the events within that hour in the timeline will be highlighted, same for the minutes and the seconds.

Examples: 
 - timeline:2007 - links to the 1st of January, 2007
 - timeline:2007-02 - links ot the 1s of February, 2007
 - timeline:2007-01-27 - links to the date of this commit
 - timeline:2007-01-27T21 - same, but with the precision set to `'hours'` (all events between 21 and 22 hours UTC will be highlighted)
 - timeline:2007-01-27T20:23 - should highlight [4646]...

Implements partially #975.

(merged from the source:sandbox/timeline-link-resolver-tmp branch, plus some fixes)

git-svn-id: https://svn.edgewall.org/repos/trac/trunk@4647 af82e41b-90c4-0310-8c96-b1721e28e2e2