Customization of the main and meta navigation entries.
The new `[mainnav]` and `[metanav]` can now be used to customize the text and link used for the navigation items, or even to disable them.

In the following example, we rename the link to the Wiki start "Home", and hide the "Help/Guide" link (#2060). We also make the "View Tickets" entry link to a specific report (#1925).

{{{
[mainnav]
wiki.label = Home
tickets.href = /report/24

[metanav]
help = disabled
}}}



git-svn-id: https://svn.edgewall.org/repos/trac/trunk@5251 af82e41b-90c4-0310-8c96-b1721e28e2e2