Added ticket cloning feature (#4686)
The ''Clone'' button is placed in the ticket box, next to the ''Reply'' one.

Note that this request uses a ''post'' method and not a ''get'' because using ''get'' doesn't work well with IE and Opera (previewing or saving a cloned ticket will fail).

Though there's no special right needed besides TICKET_CREATE, this feature is mainly useful only for TICKET_ADMINs that have to track the same issue in different line of development, so the feature will only be available to them.


git-svn-id: https://svn.edgewall.org/repos/trac/trunk@5457 af82e41b-90c4-0310-8c96-b1721e28e2e2