Improved rendering of ticket properties (merged from the FieldRefactoring branch).
The rendering of properties is now done in `_prepare_fields` in the controller rather than in the template. That method can therefore be called by `_insert_ticket_data` (for the normal ticket display) and when previewing a new ticket before submission. In the future, that method will delegate the rendering to dedicated field renderer classes and/or property renderer components.

The `_prepare_fields` method implements a few additional rendering enhancements:
 - E-mails in the CC: list are obfuscated if needed (implements #5126)
 - link to milestones (implements #449)
 - "yes/no" for checkbox custom fields instead of the "1/0" values


git-svn-id: https://svn.edgewall.org/repos/trac/trunk@5813 af82e41b-90c4-0310-8c96-b1721e28e2e2