Fixed a traceback while checking for a missing global permission 'BROWSER_VIEW' in the changeset timeline event provider.
The cause of this traceback was that a `Resource(None)` object was being created.
While this ''could'' be used for representing the fact that there's actually no real resource associated to the check, it's better to explicitly use `None` for that, as it's already done in other parts of the code.

Therefore the `Resource.__new__` method has been adjusted to take care of this special case. The documentation for that method as well as the unit tests have been expanded a bit.

git-svn-id: https://svn.edgewall.org/repos/trac/trunk@6163 af82e41b-90c4-0310-8c96-b1721e28e2e2