svn:mergeinfo renderer: optimize algorithm for computing eligible range.
We first try to get a minimal eligible range by finding the source branch revision origin, then removing the already merged revs and the blocked ones, and call `_get_node_revs` last. That costly call is restricted within the eligible range obtained so far, and is even avoided completely if the eligible range is already empty at that point.

See #8459.

git-svn-id: https://svn.edgewall.org/repos/trac/branches/0.11-stable@8401 af82e41b-90c4-0310-8c96-b1721e28e2e2