Test infrastructure: fix initial retrieval of a db connexion, right after the creation of the tables.
Doing a shutdown here flushes all the connections from the pool and therefore the next attempt to get a connection will get a new one. Only then apparently will the connection be aware of the newly created tables (remember that `init_db` can't use the connection from the pool, as it's not aware of the pool).

Should fix #8518.


git-svn-id: https://svn.edgewall.org/repos/trac/branches/0.11-stable@8416 af82e41b-90c4-0310-8c96-b1721e28e2e2