More robust rendering of the error page.
The catching of exception during evaluation of navigation contributors has been extended beyond !TracErrors, as a large range of errors can be triggered there, all likely to re-happen during rendering of the error page itself.

One category of such error are exceptions raised by a permission policy. Such an exception might be raised in any permission check, so we added a "second chance" rendering of the error page, with permission checks within the template disabled.

Closes #8585.


git-svn-id: https://svn.edgewall.org/repos/trac/trunk@8785 af82e41b-90c4-0310-8c96-b1721e28e2e2