0.11.6dev: fix two false functional test FAIL errors, related to `!#rst` rendering.
The test cases for docutils should only be added when docutils has been installed with adequate setuptools metadata, otherwise the Trac server will not be able to load the trac.mimeview.rst module due to the extra_requires constraints from setup.py.

Thanks to eli for pointing out the problem.

git-svn-id: https://svn.edgewall.org/repos/trac/branches/0.11-stable@8864 af82e41b-90c4-0310-8c96-b1721e28e2e2