WikiFormatting: make it possible to control the alignment of the content in table cells.
If the content sticks to the left border of the cell, and only to the left, then the cell will be left aligned. Conversely for the right side. If the content sticks to both sides or doesn't stick, then no cell alignment will be specified (CSS default will apply).

Example:
{{{
||align left  ||default alignment|| default alignment ||  align right||
}}}
Useful for number columns, which can be right-aligned quite naturally.


git-svn-id: https://svn.edgewall.org/repos/trac/trunk@9005 af82e41b-90c4-0310-8c96-b1721e28e2e2