Resume work on generating the API documentation (#8695).
It is now possible to build the API documentation using Sphinx via the
Makefile by issuing the following command: 
{{{
make apidoc
}}}

This will generate the HTML documentation below `build/doc/html`.
Other variations are possible, see the output of `make help` and the
documentation of [http://sphinx.pocoo.org/invocation.html sphinx-build].

Tested with Sphinx 1.0.4 on Windows (make sure `sphinx-build` is in
the PATH).

For a start, some of the testing documents have been cleaned-up, and
the documentation for the `trac.env` module is auto-generated (needs
further cleanup).



git-svn-id: https://svn.edgewall.org/repos/trac/trunk@10273 af82e41b-90c4-0310-8c96-b1721e28e2e2