0.12.6dev: re-add `msgfmt -c` to \`make check\` (from r11231)
Not all errors are caught with Python, and the msgfmt check is
performed by `tx push`. In case of a failure, it stops with the
following obscure message:
{{{
Please run this command on your system to see the errors for yourself.
}}}

Which actually means `msgfmt -c` failed.  See also the
[[http://help.transifex.com/intro/faq.html#what-should-i-do-when-my-file-has-syntax-errors| Transifex FAQ ]].


git-svn-id: https://svn.edgewall.org/repos/trac/branches/0.12-stable@11634 af82e41b-90c4-0310-8c96-b1721e28e2e2