1.0.2dev: stretch the component system so that multiple kinds of `ComponentManager` can be used.
While not excercised in Trac proper, some extensions use other
`ComponentManager` subclasses beyond the `Environment`, also in a dual
!ComponentManager/Component way.

However this breaks when both implement the same interface and the
extension point belonging to one manager tries to instantiate
the other manager, due to internal implementation details (in
particular the special case for `ComponentManager` in
`ComponentMeta.__call__`).

By introducing a limitation by which a `ComponentManager` which is
also a `Component` will only be listed in the extension points of
`Component`s managed by itself, this can be made to work.

See #11121 for more discussion.

Issue and patch by Olemis Lang <olemis+trac@gmail.com>.


git-svn-id: https://svn.edgewall.org/repos/trac/branches/1.0-stable@11758 af82e41b-90c4-0310-8c96-b1721e28e2e2