1.3.2dev: improve interoperability on Windows between native and WSL.
In a Trac installation on Windows, the paths to the repositories for
the version control sub-system are usually stored as absolute
paths. This is problematic when serving this instance using a Linux
python binary running via the Windows sub-system for Linux, as this
binary won't understand that kind of paths.

We now call `native_path` when retrieving `'dir'` key from repository
data.

The `native_path` utility knowns about Windows-like paths, and can
convert between a Windows path and a POSIX path, the target for the
conversion being Windows paths when `os.name == 'nt'` (native Windows
binaries) and POSIX paths when `os.name == 'posix'` (Linux binaries
run by the WSL).

Fixes #12653.


git-svn-id: https://svn.edgewall.org/repos/trac/trunk@15376 af82e41b-90c4-0310-8c96-b1721e28e2e2