(#12639): rework the l10n support for Jinja2 templates.
The extraction now produces "compact" messages (#11825), both for
Jinja2 `trans` directives and for the legacy Genshi templates, in
order to not have duplicate texts with different msgids in the
catalogs due to white-space difference.

Note that the `_` function is //not// producing a compact message, as
most of the time it is used for short titles or sentences.  For
multi-line content, better pick explicitly the `gettext` function,
which does the compaction (see in this changeset, file `ticket.html`).

Besides, we added support for specifying the **l10n domain** when
rendering a template, via the `'domain'` key in the metadata part of
the return value of `IRequestHandler.process_request`.

The documentation has been updated accordingly.



git-svn-id: https://svn.edgewall.org/repos/trac/trunk@15464 af82e41b-90c4-0310-8c96-b1721e28e2e2